Update hi-dpi support + finalize the next version number

+++ src/org/pushingpixels/flamingo/api/common/icon/FilteredResizableIcon.java
@@ -140,12 +140,12 @@ public class FilteredResizableIcon implements ResizableIcon {
 			BufferedImage filtered = this.operation.filter(offscreen, null);
 			this.cachedImages.put(key, filtered);
 		}
-		int scaleFactor = UIUtil.getScaleFactor();
+		double scaleFactor = UIUtil.getScaleFactor();
 		BufferedImage toDraw = this.cachedImages.get(key);
 		Graphics2D g2d = (Graphics2D) g.create();
 		g2d.translate(x, y);
-		g2d.drawImage(toDraw, 0, 0, toDraw.getWidth(null) / scaleFactor, 
-				toDraw.getHeight(null) / scaleFactor, null);
+		g2d.drawImage(toDraw, 0, 0, (int) (toDraw.getWidth(null) / scaleFactor), 
+				(int) (toDraw.getHeight(null) / scaleFactor), null);
 		g2d.dispose();
 	}
 }
